% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{An interface between the package and rstudio}
\usage{
main(path, project_type)
}
\arguments{
\item{path}{A character string. Specifies path/to/project/directory.}

\item{project_type}{The type of project to be created. See notes for
currently available project types.}
}
\value{
A new project directory with a standardised folder structure.
}
\description{
This is the function that is called when the user invokes the New Project
 wizard using the project template defined in the template file at:
 inst/rstudio/templates/project/cookiecutted.dcf
}
\note{
Currently available project types:
* Standard Project (standard_project)
* Kaggle Competition (kaggle_competition)
}
